==> notebooks/2.01-deep-final-model-trainer.ipynb <==
import dagshub
import mlflow
import os
import time
import csv
import numpy as np
import pandas as pd
import xgboost as xgb
import lightgbm as lgb
import h2o
import glob
import gc
import matplotlib.pyplot as plt
from multiprocessing import Pool
from mlflow.models.signature import infer_signature
import csv
import json
from h2o.estimators.glm import H2OGeneralizedLinearEstimator
from h2o.estimators.gbm import H2OGradientBoostingEstimator
from sklearn.metrics import roc_auc_score,precision_score,recall_score,f1_score,accuracy_score,log_loss

from instacart_model_trainer_script import ModelTrainer

os.environ['MLFLOW_TRACKING_USERNAME'] = 'tnbmarketplace'
os.environ['MLFLOW_TRACKING_PASSWORD'] = 'd7c1a451bfc46bf96fbbe5d723becf3955dfa2b3'
os.environ['MLFLOW_TRACKING_URI'] = 'https://dagshub.com/tnbmarketplace/instacart_basket_analysis_exp_tracking.mlflow'
!dagshub login --token "d7c1a451bfc46bf96fbbe5d723becf3955dfa2b3"

dagshub.init(repo_name = "instacart_basket_analysis_exp_tracking",repo_owner = 'tnbmarketplace')

# !ngrok authtoken "2kvaYw5ZiG5bL8iM8YJBVJPk1Ru_3C16mMgmpKEBYb28PPLUe"
# # Start Ngrok tunnel on H2O's port
# ngrok_tunnel = ngrok.connect(8787)
# print("H2O UI accessible at:", ngrok_tunnel)

def get_time(start):
    return time.time() - start

dataset_version = "1.2" # dataset . split method
model_version = "1.1.1" # algorithm . param version . used dataset 
read_as_xgb_dmatrix =  True
train_xgb_gbm= True
train_xgb_rf = False


train_features_name = []
with open("/kaggle/input/final-dataset-generator/train_set_columns.txt","r") as f:
    csv_reader = csv.reader(f)
    for row in csv_reader:
        train_features_name = row
train_features_name.remove('') 

test_features_name = []
with open("/kaggle/input/final-dataset-generator/test_set_columns.txt","r") as f:
    csv_reader = csv.reader(f)
    for row in csv_reader:
        test_features_name = row
test_features_name.remove('')  

train_label_index = train_features_name.index('reordered')
# test_label_index = test_features_name.index('reordered')



for i in range(len(test_features_name)):

    if test_features_name[i] == 'time_mean_dow_count':
       test_features_name[i] ='total_ord_count_p_dow'

    if test_features_name[i] == 'time_mean_ohod_count':
       test_features_name[i] ='total_ord_count_p_ohod'


xgb_booster = mlflow.xgboost.load_model("mlflow-artifacts:/c872cb0ab0ca48d481a71a9bcd2a2102/626567462ef746f8b45a4255ff51c5d1/artifacts/xgb_gbm_model")


params = {
    'max_depth': 1,
    'min_child_weight': 389,
    'verbose': -1,
    'gamma': 438,
    'eta': 0.0907183045377987,
    'subsample': 0.10741019033611729,
    'colsample_bytree': 0.336917979846298,
    'colsample_bylevel': 0.118177830385349,
    'lambda': 11,
    'alpha': 89,
    'booster': 'gbtree',
    'tree_method': 'hist',
    'objective': 'binary:logistic'
}


if read_as_xgb_dmatrix == True:
    
    train_features_name.remove('reordered')
    # test_features_name.remove('reordered')

    dtrain_2 = xgb.DMatrix(f"/kaggle/input/final-dataset-generator/final_prior_train_set.csv/part-00000-37c713df-e906-4480-9322-5a9f58652182-c000.csv?format=csv&label_column={train_label_index}",nthread=-1,feature_names=train_features_name)
    # dtest_2 = xgb.DMatrix(f"/kaggle/input/final-dataset-generator/featured_test_set.csv/part-00000-16b6977e-6b54-4aff-ad77-cb8ae65dfb2f-c000.csv?format=csv&label_column={test_label_index}",nthread=-1,feature_names=test_features_name)


if train_xgb_gbm == True:
    model_trainer = ModelTrainer("final_instacart_training",dtrain_2,params)
    # paramss = {
    #     'max_depth':1,
    #     'min_child_weight':5,
    #     'gamma':0.01,
    #     'lambda':100,
    #     'alpha':100
    # }
    xgb_gbm = model_trainer.train_xgb_gbm("c391e8337f10ceb5870cb639159539f5e3497fbf",dataset_version,model_version,params)

if train_xgb_rf == True:
    model_trainer = ModelTrainer("instacart_training",dtrain_2,dtest_2,"reordered")
    xgb_rf = model_trainer.train_xgb_rf("c391e8337f10ceb5870cb639159539f5e3497fbf",dataset_version,model_version)

    

read_as_lgb_dataset = False
train_lgb = False

if read_as_lgb_dataset == True:

    # train_features_name.remove('reordered')
#     test_features_name.remove('reordered')
    
    dtrain_2 = lgb.Dataset("/kaggle/input/final-dataset-generator/final_prior_train_set.csv/part-00000-95d78719-42e1-441c-a4e1-d662676226a3-c000.csv",params={'label_column':train_label_index},free_raw_data=False).construct()
    dtest_2 = pd.read_csv("/kaggle/input/final-dataset-generator/featured_test_set.csv/part-00000-1b691469-02d1-4ed1-8b50-ec568a4bf8a9-c000.csv",header = None)


if train_lgb == True:
    model_trainer = ModelTrainer("final_instacart_training",[dtrain_2])
    lgb_gbm = model_trainer.train_lgb_gbm("c391e8337f10ceb5870cb639159539f5e3497fbf",dataset_version,model_version)

read_as_h2o_frame = False
train_h2o = False

if read_as_h2o_frame == True:
    h2o.init(max_mem_size = '25G')
    dtrain_2 = h2o.import_file("/kaggle/input/instacart-dataset-transformer-cv2/final_prior_train_set.csv/part-00000-444aa52f-9c41-4a07-8342-7f82d9fcf6ac-c000.csv")
    dtest_2 = h2o.import_file("/kaggle/input/instacart-dataset-transformer-cv2/featured_test_set.csv/part-00000-017a6365-a37c-4ce4-8bf4-96e8daa33570-c000.csv")
    dtrain_2.col_names = train_features_name
    dtest_2.col_names = test_features_name
    dtrain_2['reordered'] = dtrain_2['reordered'].asfactor()
    dtest_2['reordered'] = dtest_2['reordered'].asfactor()

if train_h2o == True:
    
    model_trainer = ModelTrainer("instacart_training",[dtrain_2],dtest_2,"reordered")
    h2o_gbm = model_trainer.train_h2o_glm("c391e8337f10ceb5870cb639159539f5e3497fbf",dataset_version,model_version)
